Fix the current failures on the buildbot.
Patch by Kenneth Rohde Christiansen <kenneth@webkit.org> on 2009-08-11
Reviewed by Adam Treat.

As Qt hooks up the maybeDump to loadFinished, we need to make
sure that calling dump() will not call maybeDump on loadFinished.

As dump is called my emitting done() which calls dump() and then
setting m_isLoading to false. So in the case m_isLoading is false,
do not dump again.

The current code is confusing, and should be made more clear
in another commit.

* DumpRenderTree/qt/jsobjects.cpp:
(LayoutTestController::maybeDump):
(LayoutTestController::notifyDone):

Identifier: 38666@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@47065 268f45cc-cd09-0410-ab3c-d52691b4dbfc