WebCore: <rdar://problem/7141522> Incorrect layout of product table at henry.com
Reviewed by Dave Hyatt.

Test: fast/block/basic/quirk-percent-height-table-cell.html

* rendering/RenderBox.cpp:
(WebCore::RenderBox::calcPercentageHeight): Tweak the quirk that allows
percentage-height children of auto-height blocks to look for an ancestor
with non-auto height and compute their height based on it. The change is
that if that ancestor is a table cell, it is ignored and the percent
height computes to zero (just like in strict mode). This matches Firefox
and IE.

LayoutTests: Test and updated results for
<rdar://problem/7141522> Incorrect layout of product table at henry.com

Patch by Dan Bernstein <mitz@apple.com> on 2009-08-19
Reviewed by Dave Hyatt.

* fast/block/basic/quirk-percent-height-table-cell-expected.checksum: Added.
* fast/block/basic/quirk-percent-height-table-cell-expected.png: Added.
* fast/block/basic/quirk-percent-height-table-cell-expected.txt: Added.
* fast/block/basic/quirk-percent-height-table-cell.html: Added.
* fast/replaced/table-percent-height-text-controls-expected.txt:
* fast/replaced/table-percent-height-text-controls.html:



Identifier: 39118@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@47528 268f45cc-cd09-0410-ab3c-d52691b4dbfc