REGRESSION: significant slowdown on Celtic Kane "AJAX declaration" subtest https://bugs.webkit.org/show_bug.cgi?id=28332
Reviewed by Gavin Barraclough.

The method check optimisation made transitions aware of the value being
assigned when a transition was assigning a function.  This had the side
effect of making every assignment of a function expression result in a
new transition, and thus a new Structure.  The net result of this is that
the common JS idiom of

    function MyObject() {
        this.myFunction = function(...){...};
    }
    new MyObject();

Will produce a unique structure on every iteration, meaning that all
caching is defeated and there is a significant amount of structure churn.

The fix is to return the transition to its original form where it is
keyed off a property name + attributes tuple, but have each transition
support an optional transition on a specific value.

Identifier: 39187@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@47601 268f45cc-cd09-0410-ab3c-d52691b4dbfc