REGRESSION: fast/profiler/call.html is crashing occasionally https://bugs.webkit.org/show_bug.cgi?id=28476
Reviewed by Gavin Barraclough.

Using the codeblock for information about how many parameters and
locals a function has is unsafe in certain circumstances.  The
basic scenario is all function code being cleared in response to
the debugger or profiler being enabled, and then an activation is
marked before its associated function is re-executed.

To deal with this scenario we store the variable count of a function
directly in the FunctionExecutable, and then use that information.

Identifier: 39203@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@47620 268f45cc-cd09-0410-ab3c-d52691b4dbfc