Browser hangs on opening Web Inspector. https://bugs.webkit.org/show_bug.cgi?id=28438
Reviewed by Maciej Stachowiak.

Code generation needs to be able to walk the entire scopechain in some
cases, however the symbol table used by activations was a member of the
codeblock.  Following recompilation this may no longer exist, leading
to a crash or hang on lookup.

We fix this by introducing a refcounted SymbolTable subclass, SharedSymbolTable,
for the CodeBlocks used by function code.  This allows activations to
maintain ownership of a copy of the symbol table even after recompilation so
they can continue to work.

Identifier: 39210@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@47627 268f45cc-cd09-0410-ab3c-d52691b4dbfc