Fix <rdar://problem/7145242> Crashes inside WTF::HashTable below NetscapePluginInstanceProxy::disconnectStream
Reviewed by Jon Honeycutt.

* Plugins/Hosted/HostedNetscapePluginStream.mm:
(WebKit::HostedNetscapePluginStream::cancelLoad): Canceling the load can result in calling through to disconnectStream.
If the plug-in instance holds the last reference to the plug-in stream this will result in the stream being deallocated.
This leads to HostedNetscapePluginStream::cancelLoad's call to disconnectStream being passed a deallocated stream object.
Since the stream was already disconnected by the call to cancel there's no need to call disconnectStream a second time.

Identifier: 39895@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@48326 268f45cc-cd09-0410-ab3c-d52691b4dbfc