https://bugs.webkit.org/show_bug.cgi?id=27060
Patch by Norbert Leser <norbert.leser@nokia.com> on 2009-09-15
Reviewed by Darin Adler.

Symbian compiler for emulator target (WINSCW) fails with
"illegal operand" for m_attributesInPrevious in structure.ccp
(when calling make_pair functions).
This error is apparently due to the compiler not properly
resolving the unsigned type of the declared bitfield.

Initial patch explicitly casted m_attributesInPrevious
to unsigned, but since bitfield optimization is not critical for
the emulator target, this conditional change in header file
appears to be least intrusive.

* runtime/Structure.h:

Identifier: 40026@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@48461 268f45cc-cd09-0410-ab3c-d52691b4dbfc