Implement ES5 Object.defineProperty function https://bugs.webkit.org/show_bug.cgi?id=29503
Reviewed by Geoff Garen

Implement Object.defineProperty.  This requires adding the API to
ObjectConstructor, along with a helper function that implements the
ES5 internal [[ToPropertyDescriptor]] function.  It then adds
JSObject::defineOwnProperty that implements the appropriate ES5 semantics.
Currently defineOwnProperty uses a delete followed by a put to redefine
attributes of a property, clearly this is less efficient than it could be
but we can improve this if it needs to be possible in future.

Identifier: 40106@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@48542 268f45cc-cd09-0410-ab3c-d52691b4dbfc