<http://webkit.org/b/29521> run-webkit-tests: use require instead eval to load DumpRenderTreeSupport module
Reviewed by Mark Rowe.

The require statement is like the use statement, except that it
is run during script execution instead of during the 'BEGIN'
phase.  This makes it possible to change @INC before the require
statement is run.  See 'require' and 'use' in the perlfunc(1)
manpage and 'BEGIN' in perlmod(1) manpage.

* Scripts/run-webkit-tests: Replace eval statement with require
statement.

Identifier: 40133@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@48569 268f45cc-cd09-0410-ab3c-d52691b4dbfc