WebKit/mac: <rdar://problem/7290671> Crash after navigating away from a page with subframe containing plug-in
Reviewed by Brady Eidson.

At the time that -stop is called on a WebBaseNetscapePluginView that is contained within a subframe
the WebFrame's data source has been zeroed out.  This was causing -[WebBaseNetscapePluginView webFrame]
to return nil during -stop, preventing the plug-in view from unregistering itself from the PluginHalter.
We fix this by having -webFrame retrieve the frame directly rather than retrieving it from the data source.

* Plugins/WebBaseNetscapePluginView.mm:
(-[WebBaseNetscapePluginView dataSource]): Get the data source from the WebFrame.
(-[WebBaseNetscapePluginView webFrame]): Get the Frame directly from the element's document rather
than taking a windy route through the WebDataSource.

WebKitTools: Enable plug-in halting in DumpRenderTree.

We drop the plug-in halting delay to 1 second and opt in the delegate method to never halt plug-ins.
This is sufficient to ensure that the crash covered by <rdar://problem/7290671> no longer occurs.

* DumpRenderTree/mac/DumpRenderTree.mm:
(resetDefaultsToConsistentValues):
* DumpRenderTree/mac/UIDelegate.mm:
(-[UIDelegate webView:shouldHaltPlugin:]):

Identifier: 40933@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@49411 268f45cc-cd09-0410-ab3c-d52691b4dbfc