Windowed Plugins: Don't crash when client is 0.
Client is 0 when we use QWebPage without a QWebView or QGraphicsWebView.
In addition, setFrameRect()/updatePluginWidget() is called even if the
plugin was not succesfully loaded. updatePluginWidget() updates the
window rect which is, in theory, useful to draw something that indicates
that we didn't load successfully.

So, a status check is added to setNPWindowIfNeeded.

Identifier: 41225@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@49771 268f45cc-cd09-0410-ab3c-d52691b4dbfc