Improve for..in enumeration performance https://bugs.webkit.org/show_bug.cgi?id=30887
Reviewed by Geoff Garen.

Improve indexing of an object with a for..in iterator by
identifying cases where get_by_val is being used with a iterator
as the subscript and replace it with a new get_by_pname
bytecode.  get_by_pname then optimizes lookups that directly access
the base object.

Identifier: 41696@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@50254 268f45cc-cd09-0410-ab3c-d52691b4dbfc