WebKit/qt: Unreviewed buildbot fix.
Export a method to the DRT to know if the document has a
document element.

* Api/qwebframe.cpp:
(qt_drt_hasDocumentElement):

WebKitTools: Unreviewed Qt buildbot fix.

My previous fix was wrong, so revert that change and fix it by
returning when the document of the frame has no document element.
Idea is borrowed from mac and win DRT.

* DumpRenderTree/qt/DumpRenderTree.cpp:
(WebCore::DumpRenderTree::dumpFramesAsText):

Identifier: 42261@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@50832 268f45cc-cd09-0410-ab3c-d52691b4dbfc